+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Version History
 
-- 2024-??-?? 4.1.18
+- 2024-??-?? 4.1.19
+    - [#506] added robustness to Save Location change cancel
+    - [#509] fix "Load Dark" and "Load Reference"
+- 2024-12-03 4.1.18
     - deprecated RegionControlFeature
     - XL fixes
         - fixed blank wavecal